requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
sqlalchemy-serializer
